pseudo_fs: use ArcSwap to replace RwLock
PseudoFs instances should be read-most, so use ArcSwap to replace
RwLock for better performance. The ArcSwap is a linux RCU alike
technology to optimize for read-most use cases.

When ArcSwap is used, the PseudoFs.inodes and PseudoInode.children
will be copy-modify-updated, so we assume a PseudoFs instance will
not hold too much entries.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>