localfs: add an option to blob file path directly
Currently we assume that rafs blob file's name matches the blobid
string in its rafs metadata, but in somes case like rund, it's not
always the case.

For instance, a rafs blob is created with blobid "AAA" and gets
renamed to "BBB" before moving to a remote storage like nas, for now
we are not able to read any data on this rafs because of the mismatch.

Since each rafs metadata only maps to one blob, thus we can pass blob
file directly via "config.json" passed by the attach http request,
which helps rund simplify its attach logic a lot.

This adds a "blob_file" to config.json used by nydusd and
image-builder. If "blob_file" is set, the blob specified by
"blob_file" is preferred rather than the one in "dir".

Signed-off-by: Liu Bo <bo.liu@linux.alibaba.com>