builder/prefetch: also check if source root is a prefetch hint
At present, we only check if a dir is markd as hint during walk through
all inodes except root dir. If root dir is passed as a hint, it is
stored in readahead btree with None value since it is a input from
cmdline. But it has no chance to be changed to Some.

Signed-off-by: Changwei Ge <chuanlang.gcw@antfin.com>