rafs: introduce access pattern recorder
Records how a file is accessed.
For security sake, each file can associate an access pattern recorder,
which
is globally configured through nydusd configuration file.
For now, the pattern is composed of:
    1. How many times a file is read regardless of io block size and
request offset.
       And this counter can not be cleared.
    2. Last time point at which this file is read. It's wall-time in
unit of seconds.
    3. File inode number to identify which file is against.

Yes, we now don't have an abundant pattern recorder now. It can be
negotiated in the
future about how to enrich it.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>