builder: also use dev number as part of key to search inode map
A source tree  can have more than one partions mounted onto different
child directories as mountpoint. So it's possible to have two files with the same
inode number which locates in different partions. This breaks the logic
how builder tell if a file is an alias.

So fix this by use dev number as part of key.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>