rafs: also trigger prefetch according to specified list from nydusd
Prefetch files according to a directory list passed as argument when
nydusd starts in addition to prefetch table scheme persisted when
building image.

This is a dynamic method to trigger prefetch process when doing mount.
It shares most of the logic of static method based on on-disk prefetch
table.

So even with prefetch table, rafs can do prefetch now. This is good for
performance.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>