rafs: support compressed blob cache
When `compressed` option is true, use compressed blob cache such that
the local blob cache is an identical copy of the remote blob. It means
that if a chunk is compressed, it is compressed in local blob cache as
well. If it is not compressed, it is not compressed in local blob cache
too.

Signed-off-by: Peng Tao <bergwolf@hyper.sh>