fusedev: wait on epoll fd for fuse message and exit event
Let fuse service loop thread wait for epoll fd rather than blocked on
read(2). It is a feasible solution that it's not easy for us to forward
signals to each thread via pthread_kill() in Rust. Using pthread_kill to
forward signal will introduce many unsafe operations. In addition, we
can't use Rust native standard thread anymore since send a signal to it
is dangerous.

With this patch, we can easily exit from nydusd even if a process is
accessing the mountpoint since we don't rely on EBADF to exit from fuse
service loop.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>