nydusd: use cfg to control how nydusd is compiled
In fact we should not try to compile fusedev code and virtiofsd(shared)
into nydusd at the same time. We already try to do that at the cfg
section in main.rs.
Now we further apply the shceme to function `create_nydus_daemon` too.
In this way we have have different function signatures for virtiofsd and
fusedev respectively.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>