api: add APIs for Sending and Receiving fuse fd.
When nydusd works under `fusedev` mode, in order to do live upgrade, we
have to duplicate fd from nydusd to other process, usually by sending a msg
to other progress via Unix domain socket associating with a fd. In order to
notify nydusd to send and try to receive fd to and from supervisor. Open some
http Apis, thus to to interact with supervisor

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>