failover: store fuse conn id into opaque
Nydusd calculate its conn id when mount by stat(2) the mountpoint.
The conn id is just the mountpoint directory `st_dev` which is a u64
integer. Save it as part of Opaque that is backup into external
supervisor. So when do failover, new nydusd can use this conn id to find
the `flush` sysfs file to abandon unhandled fuse requests.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>

Conflicts:
	src/bin/nydusd/fusedev.rs