Builder: increase dump performance using BufWriter
Rust file I/O is unbuffered by default. If we have many small write calls
to a file, should use BufWriter. BufWriter maintains an in-memory buffer
for writing, minimizing the number of system calls required.

This patch makes image builder dump bootstrap and blob chunk using BufWriter,
so that we can increase the performance of writing a large number of small
chunks.

Need to add performance testing comparison for BufWriter in the future.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>