rafs: cap read len to length of of current VolatileSlice in copyv
The 'dst' of copyv is a slice of a VolatileSlice vector, and it may
contain multiple VolatileSlices. When writing to dst, the 'len' should
not be larger than current s.len(), otherwise s.wirte_slice() will
return error to indicate there're remaining data unwritten.

For example, reading a file larger than 4k, rafs may report the
following error

  PartialBuffer { expected: 63435, completed: 4096 }
  at rafs/src/storage/utils.rs:74

Fix it by capping the read len to the length of current VolatileSlice.

Fixes: #32448928
Signed-off-by: Eryu Guan <eguan@linux.alibaba.com>