Cargo: update reqwest to 0.11.0
reqwest depends on hyper, and prior to hyper 0.14.0 it prints the
following message with debug level, and floods log & makes debug much
harder to do.

"read 1400 bytes"

hyper 0.14.0 refactor the related code and removed the debug message.
So let's upgrade reqwest to 0.11.0, which depends on hyper 0.14.0, to
avoid this log flood.

Fixes: #34380822
Signed-off-by: Eryu Guan <eguan@linux.alibaba.com>