RAFS: add extended blob table for chunk count
Introduce a new extended blob table for RAFS format, it can be
used to record more information than the original blob table,
for example the chunk count of a blob.

Builder: for the RAFS format compatibility, we need dump extended
blob table to the tail of bootstrap.

RAFS: load blob table with extended part from bootstrap, unify the
load method for direct and cached mode.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>