rafs: don't select oss and registry backend features by default
Commit 1a3f2fe855aa ("storage: create storage lib crate") moved backend
storage into a separate crate, but made backend-oss and backend-registry
features on by default, so it's not possible for other projects to
import rafs and select only backend-localfs.

So introduce rafs/backend-oss and rafs/backend-registry features back,
so they could be selected by users. And make nydusd select them by
default.

Note that backend-localfs is now on by default for rafs, there's no need
to select it explicitly.

Fixes: 1a3f2fe855aa ("storage: create storage lib crate")
Signed-off-by: Eryu Guan <eguan@linux.alibaba.com>