Nydusify: fix possible token expires for remote request
The resolver is used for image pull or fetches requests. The best
practice in containerd is that each resolver instance is used only
once for a request and is destroyed when the request completes.

When a registry token expires, the resolver does not re-apply for
a new token, so it's better to create a new resolver instance using
resolverFunc for each request.

Signed-off-by: Yan Song <imeoer@linux.alibaba.com>