graphdriver: add docker nydus graphdriver
With nydus-snapshotter, nydus can be integrated with containerd, where
we can start/stop containers via CRI interface.

But more frequently and possibly, end-users are likely to use
docker to operate containers. So we have make this nydus graphdriver
into reality.

A limit is that it only works with oss backend rightnow, because dockerd
downloads all the layers before communicate to each kind of grapdriver.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>