snapshotter: support oss and localfs backend as well
Currently, nydus snapshotter assumes the backend is registry, as
NewDaemonConfig() doesn't check backend type and always set 'Host',
'Repo' and 'Auth' fields that are only useful for registry backend.

So add oss and localfs backend configs to Backend config as well, so
images with such backends could be used correctly.

Fixes: #35317323
Signed-off-by: Eryu Guan <eguan@linux.alibaba.com>