blobcache: don't check chunk existence before insert it
Let blobcache itself to check if chunks is ready/up-to-date
before modify the chunks map.
Basically, blobcache business should not try to `set_ready`
frequently. The method `set_ready` usually should be called once.

So in theory, we can save some cpu cycles with this patch.

Moreover, we don't have to worry about inserting confliction as the value
is always a true value.

Signed-off-by: Changwei Ge <chge@linux.alibaba.com>